+++ src/main/java/net/silentchaos512/mechanisms/block/batterybox/BatteryBoxScreen.java
@@ -47,10 +47,10 @@ public class BatteryBoxScreen extends ContainerScreen<BatteryBoxContainer> {
         }
 
         // Debug text
-        int y = 5;
-        for (String line : container.tileEntity.getDebugText()) {
-            font.drawString(line, 5, y, 0xFFFFFF);
-            y += 10;
-        }
+//        int y = 5;
+//        for (String line : container.tileEntity.getDebugText()) {
+//            font.drawString(line, 5, y, 0xFFFFFF);
+//            y += 10;
+//        }
     }
 }