+++ src/main/java/net/silentchaos512/mechanisms/block/electricfurnace/ElectricFurnaceTileEntity.java
@@ -21,7 +21,7 @@ public class ElectricFurnaceTileEntity extends AbstractMachineTileEntity<Abstrac
     // Energy constant
     public static final int MAX_ENERGY = 50_000;
     public static final int MAX_RECEIVE = 500;
-    public static final int ENERGY_USED_PER_TICK = 25;
+    public static final int ENERGY_USED_PER_TICK = 30;
 
     // Inventory constants
     private static final int INVENTORY_SIZE = 2;