+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Ore chunks, dusts, and ingots for copper, tin, silver, lead, and nickel. This does NOT add ores! The recipes will load if another mod defines an ore with the correct tag.
+## Removed
+- Debug text from coal generator and battery box GUI
+### Changed
+- Coal generator now creates 60 FE/tick (from 25)
+- Machines now use 30 FE/tick (from 25)
 
 ## [0.1.0] - 2019-07-06
 - First release
-